\name{Package version}
\alias{version.qtlDesign}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Version of qtlDesign package}
\description{
Returns the version number for the qtlDesign package.}
\usage{
version.qtlDesign()
}
%- maybe also 'usage' for other objects documented here.

\value{The version number.
}

\author{Saunak Sen, Jaya Satagopan, Karl Broman, and Gary
  Churchill}
\keyword{design}
