\name{qtlDesign-internal}
\alias{qtlDesign-internal}
\alias{nu}
\alias{tau}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{qtlDesign-internal}
\description{
Internal qtlDesign functions
}
%\usage{}
%\arguments{}
\details{
  These are not to be called by the user.
}

\value{Scalar value of the nu and tau functions in Siegmund (1985).
}

\references{Siegmund, D., 1985 Sequential Analysis: Tests and Confidence
  Intervals.  Springer-Verlag, New York.}
\author{Saunak Sen}

%\examples{}
\keyword{internal}
