\name{qtlDesign-internal}
\alias{qtlDesign-internal}
\alias{nu}
\alias{tau}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{qtlDesign-internal}
\description{
Internal qtlDesign functions
}
%\usage{}
%\arguments{}
\details{
  These are not to be called by the user.
}

\references{Sen S, Satagopan JM, Churchill GA (2005) 	
Quantitative trait locus study design from an information perspective.
Genetics, 170:447-64.}
\author{Saunak Sen, Jaya Satagopan, Karl Broman, and Gary Churchill}

%\examples{}
\keyword{internal}
